+++ Display/ScrollingMessages.lua
@@ -9,6 +9,7 @@ addonTable.Display.ScrollingMessagesMixin = {}
 function addonTable.Display.ScrollingMessagesMixin:MyOnLoad()
   self:SetHyperlinkPropagateToParent(true)
   self:SetClipsChildren(true)
+  self:SetFlattensRenderLayers(true)
 
   self.scrollIndex = 1
 
@@ -20,6 +21,16 @@ function addonTable.Display.ScrollingMessagesMixin:MyOnLoad()
 
   self.pool = CreateFontStringPool(self, "BACKGROUND", 0, addonTable.Messages.font)
   self.barPool = CreateTexturePool(self, "BACKGROUND")
+
+  do
+    local edgeFadeTop = 20
+    local edgeFadeLeft = 0
+    local edgeFadeRight = 0
+    local edgeFadeBottom = 0
+
+    self:SetAlphaGradient(0, CreateVector2D(edgeFadeLeft, edgeFadeTop))
+    self:SetAlphaGradient(1, CreateVector2D(edgeFadeRight, edgeFadeBottom))
+  end
 
   self:SetScript("OnMouseWheel", function(_, delta)
     self.currentFadeOffsetTime = GetTime()
@@ -227,7 +238,7 @@ function addonTable.Display.ScrollingMessagesMixin:Render(newMessages)
         fs:SetJustifyH("LEFT")
         fs:SetPoint("LEFT", self, addonTable.Messages.inset + 3, 0)
         fs:SetPoint("RIGHT", self)
-        fs:SetPoint("BOTTOM", self)
+        fs:SetPoint("BOTTOM", self, 0, 2)
         fs:SetText(m.text)
         fs:SetTextColor(m.color.r, m.color.g, m.color.b)
         fs:SetTextScale(addonTable.Messages.scalingFactor)